// src/app/models/referans.ts
export interface AnaRaporTuruResponse {
  anaRaporTuruId: string;
  anaRaporTuruKodu: string;
  anaRaporTuruAdi: string;
}

export interface RaporTuruResponse {
  raporTuruId: string;
  raporTuruKodu: string;
  raporTuruAdi: string;
}

export interface VergiKoduResponse {
  vergiKoduId: string;
  vergiKodu: string;
  vergiKoduAdi: string;
}